Stock-Trading-Advisor/requirements.txt
ansicon==1.89.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
awscli==1.22.49
awsebcli==3.20.3
bcrypt==3.2.0
blessed==1.19.1
botocore==1.23.49
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.3
cryptography==36.0.1
databases==0.5.5
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
ecdsa==0.17.0
fastapi==0.73.0
future==0.16.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
jinxed==1.1.0
jmespath==0.10.0
jsonschema==3.2.0
paramiko==2.9.2
passlib==1.7.4
pathspec==0.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.1
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
watchgod==0.7
wcwidth==0.1.9
websocket-client==0.59.0
websockets==10.2
